requirements.txt
blinker==1.6.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.7
decorator==4.4.2
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
idna==3.4
imageio==2.31.3
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
moviepy==1.0.3
numpy==1.26.0
pika==1.3.2
Pillow==10.0.1
proglog==0.1.10
pycparser==2.21
python-dotenv==1.0.0
pytube==15.0.0
requests==2.31.0
SQLAlchemy==2.0.21
srt==3.5.3
tqdm==4.66.1
typing_extensions==4.8.0
urllib3==2.0.4
vosk==0.3.45
websockets==11.0.3
Werkzeug==3.0.0
